Move cli args to main
